#54 [Refactor] 스타일 및 토스트 문제 해결
